\name{BoostingLVM-package}
\alias{BoostingLVM-package}
\alias{BoostingLVM}
\docType{package}
\title{
\packageTitle{BoostingLVM}
}
\description{
\packageDescription{BoostingLVM}
}
\details{
Function \code{\link{lvmboost}} fits a factor analysis model for binary data
by likelihood-based boosting. Function \code{\link{plot.lvmboost}}
plots the regularization path and function \code{\link{cv.lvmboost}}
selects the number of iterations by cross-validation.
A simulated dataset can be obtained using function \code{\link{simdatalvm}}.
}
\author{
\packageAuthor{BoostingLVM}

Maintainer: \packageMaintainer{BoostingLVM}
}
\references{
}
