\name{plot.lvmboost}
\alias{plot.lvmboost}
\title{Regularization plot}
\description{
Regularization plot of the latent variable model fitted by boosting
}
\usage{
\method{plot}{lvmboost}(x, which = 0, ...)
}
\arguments{
\item{x}{an object of class \code{lvmboost} obtained from function \code{\link{lvmboost}}.}
\item{which}{vector of parameters to plot. If 0 all the parameters are plot.}
\item{\dots}{further arguments passed to or from other methods.}
}
%\references{
%}
\seealso{
\code{\link{lvmboost}}
}
\examples{}
\author{}

